refactor: Move components to feature directory
